Refactor large files into smaller components and extract logic.
Refactored `Tree.tsx` into `TreeBranches.tsx`, `TreeLeaves.tsx`, `TreeFruits.tsx`, and `Tree.tsx`.
Extracted `createBranchGeometry` to `geometry.ts`.
Refactored `GenomePanel.tsx` by extracting `RuleCard` into `src/app/components/panels/genome/RuleCard.tsx`.
